ean - moved flask application up one level
