+++ porthole/gui/__init__.py
@@ -1,1 +1,2 @@
 from .app import edit_config, add_connection, delete_connection, read_config, ConnectionForm, ConfigForm, config
+# from ..connections import ConnectionManager
